requirements.txt
flask==2.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
werkzeug==2.2.3